Final modulo operations CLI
Tested as everything works well. Interchangeable blending when using even amount of layers, but treat alpha as if it was a blending mode operation or you can apply manipulate value by a modulo value with a formula. Included the formula as a separate formula for usage outside of batch processing for easier filter writing.