Replaced obsolete filter reduced color transfer
Can't guarantee that it'll work 100%, but have tested, and checked many time to make sure no bugs will likely be found nor errors from manual copy and pasting due to the lack of having linked file (yes, tried it, doesn't work regardless of what I put in, but that could be me). Not going to bother getting pal i or pal -1 check for duplicate colors, that's up to the users, as it's much more work than I want to put in for now, I'll do that later. That being said, It seems the filter does what it suppose to do. 

pal_t has $8, and $9 reverted for easier coding and cli application.