Finally made the mosaic work!
The reason why it was resizing is because iw, and ih was taking the last image dimension rather than the first image which is the target image. Setting them to w#0, and h#0 didn't fix my problem. So, my final resort is to make the input_glob, and file location as part of the command and it worked.

New feature has been added - Disable autocrop to original dimension or enable it. 

Furthermore, the reason why the gui filter wasn't working is that I wasn't quite reading things right. I forgot the '_' and I did added a variable above the preview code to make sure that it works though I don't know if that was necessary, but it's best to make sure anyway.

I find that there might be complications with gui split preview using default code. That will have to be done later though I'm quite content with the current implementation. I added a note about split preview. Everything works as expected though sometimes there might be issues there and there with autocrop to median off, turning it on fixes issues at times, so that's what it's there for as it may or may not be needed to get things to work.