Major upgrade to Fragment Blur
After realizing about using the +channels {s-1} and using that to indicate number of alpha layers to solve my problem, I was able to finish this filter. With using that, I was able to make the Fragment Blur match PDN behavior 100% of the time. Now, here's the major upgrade to Fragment Blur. 

1. Speed is over 500 times faster though not as fast as PDN version, but it's usable in many more cases. From a test I did with 50 iteration with a image that has 4000x3000. It goes from 30-90 minutes to 4 seconds. That's how fast it is now.
2. Better handling of picture info. If some info are not needed to be processed, they will not be processed. In some cases, some options are taken out because it doesn't match expectations.
3. Behaves exactly like the Paint.NET version