Added 2 new cli commands, and CLI fixes
The new commands are related to perspective distortion, and it is based of rectangular polar coordinates. 

What's next is for me to figure out how to create lines within axis taking into account of opacity, and using existing pixels. This will be applied between the transformations to finally recreate perspective with vanishing point.