valuesmanip:          Directory for tutorial pages concerning the                       Values Manipulation category.
valuesmanip/fill.gmd: Refurbished tutorial on the fill command.

This tutorial has been shortened from the original, as the text in the
second half of the original tutorial developed material better
situated for a math scripting tutorial. As is, it is still a long
tutorial and it offered something of a stress test of the new
markdown tools. Some cheating here. For precise image sizing, used
the "%%  %%" notation, which invoked `gmic` for exact image sizing.
The Markdown notation ![alt](URL) is probably cheaper, and suits
in the majority of cases where a precise image size is not needed.