valuesmanip/orientation.gmd: Refurbished tutorial on the index command.
Orientation has been adapted from original, but the examples have
changed, with a beginner-to-intermediate inclusion of a math
expression to illustrate how data produced by -orientation might be
utilized. I plan other such incursions of math expressions in other
tutorials; given their significance, I'm not as inclined to hold them
at arm's length as I once did. A math expression tutorial is in the
offing, some day, but offloading some math expression tutorial
material to this line of tutorials is probably a good idea, and leads
to greater integration of the two worlds in the minds of G'MIC
students. The key-color example of the original tutorial can be
adapted to use -norm and placed there. -norm and -orientation are two
aspects of vectors, so it is appropriate to put key-color there.