modified:   tutorial/Beginner_s_Cookbook/Arabesque.gmd modified:   tutorial/Beginner_s_Cookbook/images/phase.mp4 new file:   tutorial/Beginner_s_Cookbook/images/phase_sum.mp4 new file:   tutorial/Beginner_s_Cookbook/images/wheeliechainmap.svg new file:   tutorial/Beginner_s_Cookbook/images/wheeliepaths.svg modified:   tutorial/Beginner_s_Cookbook/scripts/wheelie.gmic modified:   tutorial/Beginner_s_Cookbook/scripts/wheelie_anim.gmic modified:   tutorial/Beginner_s_Cookbook/scripts/wheelie_simple.gmic
A replacement of 'Demon's Arise', 'Wheelie Math' and subsections furnish the background
theory on how to find wheelies that draw ampersands and other pre-conceived
arabesques. It develops the underlying ideas of the discrete time Discrete
Fourier Transform and its inverse, but in wheelie terms: what is a complex
exponential term? How does a sum of complex exponential terms differ from
wheelie chains and the many-to-one relationship between all the wheelie chains
that draw a given arabesque, K, and the one complex exponential sum that
defines K. This culminates in, but does not pose, performing a discrete time
Fourier Transform to get a spectrum (aka, complex exponential sum) of an
arbitrarily drawn closed shaped - someone's face, perhaps, and from that
choosing a particular permutation to assemble a wheelie chain that draws
the shape. This is the consolidated commit of the 'Vanishing' branch, now
incorporated back into master.