Cleanup: # Wed Aug 3 2022 - Mon Sep 5 2022
Revised for a more direct exposition, so that the chains of wheelies
first described in "Wheelies" are identified as "delta
wheelies". There is an underlying "canonical sum" of wheelies that
define an arabesque and which constitute an arabesque's spectral
components.  Delta wheelies — these rotating line segments
identifiable in arabesque-drawing animations — emerge from taking
pairwise orientation and rotational differences between terms of the
canonical sum. Pairing depends on ordering, so chains of delta
wheelies in arabesque drawing arrangements are commensurate with
particular orderings, or permutations, of the terms of the canonical
sum. The wheelies composing this sum do not compare with any chain
drawing set, but through differencing can generate each and every one
of them. The canonical sum defines the arabesque that each chain draws
and the terms of the canonical sum specifically compare to the
spectral content of the arabesque. Regarding an arabesque as a time
series and taking its discrete time Fourier transform directly produce
the terms of the canonical sum. Evaluating that cononical sum over
time compares to taking the discrete inverse Fourier transform.

modified:   Arabesque.gmd
	    images/wheelmult.mp4
	    images/wheeliechainmap.svg
            scripts/spectralarabesque.gmic
            scripts/wheelie_simple.gmic
new:        images/dual.mp4

modified:   ../../include/tom_keil.gmic
            A patch for fx_tk_autodepth, which — especially for small images — would transform
            the height of its selected images differently than the width, breaking pipelines
            where unvarying image dimensions are assumed. The break was due to rounding errors
            that changed image dimensions. Addressed by changing strict conformance from
            preserving aspect ratios to preserving image dimensions.