Added hashmap processing into rep_mode_blend After spending a lot of time getting the dictionary approach to work. I tested it, and this is my findings:
Test 1: 13 Images of 256x256
   With hashmap - 0.053 s
   Without hashmap - 0.016 s
Test 2: 201 Images of 650x650
   With hashmap - 3.5 s
   Without hashmap 8.6 s

So, I went further and test the amount of images to switch between mode. So, if you have less than 20 images, and assuming you know the images have duplicates, it will be faster than the dictionary approach. Otherwise, it switches to the dictionary approach.